backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pyodbc==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
lxml==4.9.3
fake-useragent==1.4.0
schedule==1.2.0
dateparser==1.2.0
gunicorn==21.2.0
waitress==2.1.2